+++ src/App.tsx
@@ -1,8 +1,7 @@
 import React  from 'react';
 import './App.css'; 
 import TimeTable from './components/TimeTable';
-
-
+ 
 const App: React.FC = () => {
  
   return (
