+++ src/index.tsx
@@ -2,18 +2,10 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
-import * as serviceWorker from './serviceWorker';
-import {
-  ApolloClient, ApolloProvider, HttpLink, InMemoryCache
-} from '@apollo/client'
+import { ApolloProvider } from '@apollo/client'
+import { client } from './graphQL'
 
-const client = new ApolloClient({
-  cache: new InMemoryCache(),
-  link: new HttpLink({
-    uri: 'https://api.digitransit.fi/routing/v1/routers/hsl/index/graphql',
-  })
-})
-
+ 
 ReactDOM.render(
   <React.StrictMode>
     <ApolloProvider client={client}>
@@ -21,9 +13,4 @@ ReactDOM.render(
     </ApolloProvider>
   </React.StrictMode>,
   document.getElementById('root')
-);
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
+);
