--- src/utils/hooks/index.ts
@@ -1,20 +1,0 @@
-import { useEffect, useRef } from 'react';
-
-export const useInterval = (callback: () => void, delay: number) => {
-    const savedCallback: any = useRef();
-
-    useEffect(() => {
-        savedCallback.current = callback;
-    }, [callback]);
-
-    useEffect(() => {
-        function tick() {
-            savedCallback.current();
-        }
-        if (delay !== null) {
-            let id = setInterval(tick, delay);
-            return () => clearInterval(id);
-        }
-    }, [delay]);
-};
-
